+++ asset_allocation/risk_management/high_to_low.py
@@ -19,8 +19,6 @@ class HighToLow(stop_loss.StopLoss):
         frequency of return data
     rebelance: str
         rebalance frequency
-    rebalance_dates: list
-        list of rebalancing dates
     """
 
     def __init__(
@@ -29,12 +27,9 @@ class HighToLow(stop_loss.StopLoss):
         stop_threshold: float,
         frequency: str,
         rebalance: str,
-        rebalance_dates: list,
         **kwargs,
     ) -> None:
-        super().__init__(
-            universe, stop_threshold, frequency, rebalance, rebalance_dates, **kwargs
-        )
+        super().__init__(universe, stop_threshold, frequency, rebalance, **kwargs)
         self.highs = np.zeros(shape=self.num_total_assets)
 
     def assign(