+++ mathstats/mean/simple_mean.py
@@ -144,3 +144,14 @@ class SimpleMean(object):
             batch_weight=batch_weight,
             **kwargs,
         )
+
+    def is_valid(self):
+        """
+        check if inputs are valid
+
+        Returns
+        -------
+        bool
+            True if inputs are valid, false otherwise
+        """
+        return self.total_iterations > 0