+++ CHANGELOG.md
@@ -8,8 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Original Weight Weighting scheme
+- Add Is Valid validation in risk and return engines
 ### Fixed
+- Fixed Monthly vs. Daily Bug in Covariance Matrix
 ### Changed
+- Only update return and risk engine once, not in every strategy
 ### Removed
 
 ## [1.5.1] - 2023-12-18